chore: update Cypress config file and move skipped test out of basic dir [PT-186622946]
Cypress has a bug that breaks tests that import files imported from
outside the Cypress folder when TypeScript 5+ is used:

https://github.com/cypress-io/cypress/issues/26203

These tests were currently skipped anyway, so moving them out of the
basic dir so Cypress doesn't even try to run them and fail because
of the sourceMap issue. Hopefully, this Cypress bug gets fixed
before we will have time to re-enable these tests.
