+++ packages/tectonic-explorer/cypress.config.ts
@@ -17,6 +17,6 @@ export default defineConfig({
     },
     baseUrl: 'http://localhost:8080',
     specPattern: 'cypress/e2e/**/*.{js,jsx,ts,tsx}',
-    testIsolation: 'strict',
+    testIsolation: false
   },
 });
