+++ packages/tectonic-explorer/cypress/e2e/simulation-rules/continent-ocean-collision.js
@@ -1,7 +1,7 @@
-import Model from "../../../../src/plates-model/model";
-import presets from "../../../../src/presets";
-import * as modelHelper from "../../../../src/plates-model/model-test-helpers";
-import { getImageData } from "../../../../src/utils";
+import Model from "../../../src/plates-model/model";
+import presets from "../../../src/presets";
+import * as modelHelper from "../../../src/plates-model/model-test-helpers";
+import { getImageData } from "../../../src/utils";
 
 let modelImgData = null;
