+++ packages/tectonic-explorer/cypress/e2e/simulation-rules/continental-collision.js
