+++ packages/tectonic-explorer/cypress/e2e/simulation-rules/island-collision.js
